+++ packages/sdk-js/src/DidHelpers/createDid.ts
@@ -5,55 +5,19 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-import {
-  DidVerificationMethodType,
-  getStoreTx,
-  keypairToMultibaseKey,
-  multibaseKeyToDidKey,
-  signingMethodTypes,
-} from '@kiltprotocol/did'
-import { KiltAddress, KeyringPair, SignerInterface } from '@kiltprotocol/types'
-import * as Did from '@kiltprotocol/did'
-import type {
-  DidDocument,
-  DidUrl,
-  VerificationMethod,
-} from '@kiltprotocol/types'
-import { Signers } from '@kiltprotocol/utils'
+import { getFullDid, getStoreTx, signingMethodTypes } from '@kiltprotocol/did'
+import type { KiltAddress, SignerInterface } from '@kiltprotocol/types'
+import { Crypto, Signers } from '@kiltprotocol/utils'
+
+import { checkResultImpl, convertPublicKey, submitImpl } from './common.js'
 import type {
   AcceptedPublicKeyEncodings,
-  KeyMultibaseEncoded,
   SharedArguments,
   TransactionHandlers,
 } from './interfaces.js'
-import { transact } from './index.js'
 
-export function convertPublicKey(pk: AcceptedPublicKeyEncodings): {
-  publicKey: Uint8Array
-  keyType: string
-} {
-  let didPublicKey: {
-    publicKey: Uint8Array
-    keyType: string
-  }
-  if (typeof pk === 'string') {
-    didPublicKey = multibaseKeyToDidKey(pk)
-  } else if ('publicKeyMultibase' in pk) {
-    didPublicKey = multibaseKeyToDidKey(
-      (pk as { publicKeyMultibase: KeyMultibaseEncoded }).publicKeyMultibase
-    )
-  } else if (
-    'publicKey' in pk &&
-    pk.publicKey.constructor.name === 'Uint8Array'
-  ) {
-    didPublicKey = {
-      publicKey: pk.publicKey,
-      keyType: pk.type,
-    }
-  } else {
-    throw new Error('invalid public key')
-  }
-  return didPublicKey
+function implementsSignerInterface(input: any): input is SignerInterface {
+  return 'algorithm' in input && 'id' in input && 'sign' in input
 }
 
 /**
@@ -62,87 +26,74 @@ export function convertPublicKey(pk: AcceptedPublicKeyEncodings): {
  * @param options.fromPublicKey The public key that will feature as the DID's initial authentication method and will determine the DID identifier.
  * @param options
  */
-export async function createDid(
+export function createDid(
   options: Omit<SharedArguments, 'didDocument'> & {
     fromPublicKey: AcceptedPublicKeyEncodings
   }
-): Promise<TransactionHandlers> {
-  const didPublicKey = convertPublicKey(options.fromPublicKey)
+): TransactionHandlers {
+  const getSubmittable: TransactionHandlers['getSubmittable'] = async (
+    submitOptions = {}
+  ) => {
+    const { fromPublicKey, submitter, signers, api } = options
+    const { signSubmittable = true } = submitOptions
+    const { publicKey, keyType } = convertPublicKey(fromPublicKey)
 
-  if (!signingMethodTypes.includes(didPublicKey.keyType)) {
-    throw new Error('invalid public key')
-  }
+    if (!signingMethodTypes.includes(keyType)) {
+      throw new Error('invalid public key')
+    }
+    const submitterAccount = (
+      'address' in submitter ? submitter.address : submitter.id
+    ) as KiltAddress
 
-  const didPublicKeyMultibase = keypairToMultibaseKey({
-    publicKey: didPublicKey.publicKey,
-    type: didPublicKey.keyType as DidVerificationMethodType,
-  })
-
-  const submitterAccount = (
-    'address' in options.submitter
-      ? options.submitter.address
-      : options.submitter.id
-  ) as KiltAddress
-
-  const didAuthKey: Did.NewDidVerificationKey = {
-    publicKey: didPublicKey.publicKey,
-    type: didPublicKey.keyType as any,
-  }
-
-  function implementsSignerInterface(input: any): boolean {
-    return 'algorithm' in input && 'id' in input && 'sign' in input
-  }
-
-  const signers: SignerInterface[] = (
-    await Promise.all(
-      options.signers
-        .map(async (signer) => {
+    const accountSigners = (
+      await Promise.all(
+        signers.map(async (signer) => {
           if (implementsSignerInterface(signer)) {
-            return signer as SignerInterface
+            return [signer]
           }
-          const res = (await Signers.getSignersForKeypair({
-            keypair: signer as
-              | KeyringPair
-              | {
-                  secretKeyMultibase: KeyMultibaseEncoded
-                  publicKeyMultibase: KeyMultibaseEncoded
-                },
-          })) as SignerInterface[]
+          const res = await Signers.getSignersForKeypair({
+            keypair: signer,
+          })
           return res
         })
-        .flat()
+      )
+    ).flat()
+    const didCreation = await getStoreTx(
+      {
+        authentication: [{ publicKey, type: keyType as 'sr25519' }],
+      },
+      submitterAccount,
+      accountSigners
     )
-  ).flat()
 
-  const fullDid = Did.getFullDidFromVerificationMethod({
-    publicKeyMultibase: didPublicKeyMultibase,
-  })
+    let signedHex
+    if (signSubmittable) {
+      const signed =
+        'address' in submitter
+          ? await didCreation.signAsync(submitter)
+          : await didCreation.signAsync(submitterAccount, {
+              signer: Signers.getPolkadotSigner([submitter]),
+            })
+      signedHex = signed.toHex()
+    } else {
+      signedHex = didCreation.toHex()
+    }
 
-  const vm: VerificationMethod = {
-    id: `${fullDid}#auth` as DidUrl,
-    controller: fullDid,
-    type: 'Multikey',
-    publicKeyMultibase: didPublicKeyMultibase,
+    return {
+      txHex: signedHex,
+      checkResult: (input) =>
+        checkResultImpl(
+          input,
+          api,
+          [{ section: 'did', method: 'DidCreated' }],
+          getFullDid(Crypto.encodeAddress(publicKey, 38)),
+          signers
+        ),
+    }
   }
 
-  const didDocument: DidDocument = {
-    id: fullDid,
-    verificationMethod: [vm],
-    authentication: [vm.id],
-  }
+  const submit: TransactionHandlers['submit'] = (submitOptions) =>
+    submitImpl(getSubmittable, { ...options, ...submitOptions })
 
-  const fullDidCreationTx = await getStoreTx(
-    {
-      authentication: [didAuthKey],
-    },
-    submitterAccount,
-    signers
-  )
-
-  return transact({
-    ...options,
-    didDocument,
-    call: fullDidCreationTx,
-    expectedEvents: [{ section: 'did', method: 'DidCreated' }],
-  })
+  return { getSubmittable, submit }
 }
